Mais testes do repositório

+++ src/main/java/com/cotemig/repositories/FeeRepository.java
@@ -3,6 +3,6 @@ package com.cotemig.repositories;
 import com.cotemig.models.Fee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FeeRepository extends JpaRepository<Fee, Long> {
+public interface FeeRepository extends JpaRepository<Fee, Integer> {
 
 }
